Moved methods from TFSWorldLoader.TreeParsing to TFSWorldLoader. Simplified Tile's implementation.
