+++ COMMO.GameServer/World/Loading/TFSWorldLoader.cs
@@ -1,5 +1,6 @@
 using COMMO.GameServer.OTBParsing;
 using System;
+using System.Collections.Generic;
 
 namespace COMMO.GameServer.World.Loading {
 
@@ -28,7 +29,81 @@ namespace COMMO.GameServer.World.Loading {
 			}
 
 			throw new NotImplementedException();
+		}
 
+		public static TFSWorldHeader GetWorldHeader(OTBTree tree) {
+			if (tree == null)
+				throw new ArgumentNullException(nameof(tree));
+
+			var parsingStream = new OTBNodeParsingStream(tree, tree.Root);
+
+			var worldEncodingVersion = parsingStream.ReadUInt32();
+			var worldWidth = parsingStream.ReadUInt16();
+			var worldHeight = parsingStream.ReadUInt16();
+			var itemEncodingMajorVersion = parsingStream.ReadUInt32();
+			var itemEncodingMinorVersion = parsingStream.ReadUInt32();
+
+			return new TFSWorldHeader(
+				worldEncodingVersion: worldEncodingVersion,
+				worldWidth: worldWidth,
+				worldHeight: worldHeight,
+				itemEncodingMajorVersion: itemEncodingMajorVersion,
+				itemEncodingMinorVersion: itemEncodingMinorVersion);
+		}
+
+		public static TFSWorldAttributes GetWorldAttributes(OTBTree tree) {
+			if (tree == null)
+				throw new ArgumentNullException(nameof(tree));
+
+			if (tree.Root.Children.Count != 1)
+				throw new MalformedWorldException();
+
+			var worldDataNode = tree.Root.Children[0];
+			if ((OTBNodeType)worldDataNode.Type != OTBNodeType.WorldData)
+				throw new MalformedWorldException();
+
+			var parsingStream = new OTBNodeParsingStream(tree, worldDataNode);
+
+			var worldDescription = new List<string>();
+			string spawnsFilename = null;
+			string housesFilename = null;
+
+			while (!parsingStream.IsOver) {
+				var attribute = (TFSWorldNodeAttribute)parsingStream.ReadByte();
+				switch (attribute) {
+					case TFSWorldNodeAttribute.WorldDescription:
+					worldDescription.Add(parsingStream.ReadString());
+					break;
+
+					case TFSWorldNodeAttribute.ExtensionFileForSpawns:
+					if (spawnsFilename != null) {
+						throw new MalformedWorldAttributesNodeException("Multiple filenames for world spawns.");
+					} else {
+						spawnsFilename = parsingStream.ReadString();
+					}
+					break;
+
+					case TFSWorldNodeAttribute.ExtensionFileForHouses:
+					if (housesFilename != null) {
+						throw new MalformedWorldAttributesNodeException("Multiple filenames for world houses.");
+					} else {
+						housesFilename = parsingStream.ReadString();
+					}
+					break;
+
+					default:
+					throw new MalformedWorldAttributesNodeException("Unknown attribute found in world attributes note.");
+				}
+			}
+
+			var formattedWorldDescription = string.Join(
+				separator: Environment.NewLine,
+				values: worldDescription);
+
+			return new TFSWorldAttributes(
+				worldDescription: formattedWorldDescription,
+				spawnsFilename: spawnsFilename,
+				housesFilename: housesFilename);
 		}
 	}
 }
