+++ COMMO.GameServer/World/Tile.cs
@@ -4,20 +4,23 @@ using COMMO.GameServer.Items;
 
 namespace COMMO.GameServer.World {
 
-    public class Tile {
-        public readonly ushort X;
-        public readonly ushort Y;
-        public readonly byte Z;
+	/// <summary>
+	/// This class represents a tile in the world.
+	/// It's position is described using 2 ushort and a byte because
+	/// the it's a relative position within a "floor"
+	/// </summary>
+    public sealed class Tile {
+		public readonly Position Position;
+		public readonly bool BelongsToHouse;
 
         public readonly List<Item> Items = new List<Item>();
         public readonly List<PlayerCharacter> PlayerCharacters = new List<PlayerCharacter>();
 
         public TileFlags Flags;
 
-        public Tile(ushort x, ushort y, byte z) {
-			X = x;
-			Y = y;
-			Z = z;
+        public Tile(Position position, bool belongsToHouse) {
+			Position = position;
+			BelongsToHouse = belongsToHouse;
         }
 
         public void AddInternalThing(Item item) {
